+++ src/inc/api/APISendProgress.class.php
@@ -68,6 +68,10 @@ class APISendProgress extends APIBasic {
       $this->sendErrorResponse(PActions::SEND_PROGRESS, "The given task does not have a corresponding hashlist!");
     }
     $totalHashlist = $hashlist;
+    $isSuperhashlist = false;
+    if ($hashlist->getFormat() == DHashlistFormat::SUPERHASHLIST) {
+      $isSuperhashlist = true;
+    }
     $hashlists = Util::checkSuperHashlist($hashlist);
     foreach ($hashlists as $hashlist) {
       if ($hashlist->getIsSecret() > $this->agent->getIsTrusted()) {
@@ -298,6 +302,11 @@ class APISendProgress extends APIBasic {
     }
     $chunk = Factory::getChunkFactory()->get($chunk->getId());
     $chunk->setCracked($chunk->getCracked() + $sumCracked);
+    if ($isSuperhashlist) {
+      // if it's a superhashlist, we need to update the count for the superhashlist as well
+      $hashlist = Factory::getHashlistFactory()->get($taskWrapper->getHashlistId());
+      Factory::getHashlistFactory()->inc($hashlist, Hashlist::CRACKED, $sumCracked);
+    }
     Factory::getChunkFactory()->update($chunk);
     Factory::getAgentFactory()->getDB()->commit();